+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - rename (fix typo) argument to `lcs_child_in_parent` in `CoordinateSystemManager.add_cs`  \[{pull}`936`\].
 - replace usages of `pkg_resources` with `importlib.metadata` \[{pull}`941`\].
 - replace usages of `copy_arrays` with `memmap` for `asdf>=3.1.0` \[{pull}`940`\].
+- added support for homogeneous transformation matrices:
+  - added `create_cs_from_homogenous_transformation` to `CoordinateSystemManager`
+  - added `from_homogeneous_transformation` to `LocalCoordinateSystem`
+  - added `as_homogeneous_matrix` to `LocalCoordinateSystem`
 
 ### Dependencies